requirements.txt
tensorflow==2.7.0
tensorflow-hub==0.12.0
numpy
matplotlib
streamlit==1.6.0
Pillow==9.0.1
opencv-python==4.5.5.62